requirements.txt
altair==5.0.1
attrs==23.1.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.4
decorator==5.1.1
filelock==3.12.2
gitdb==4.0.10
GitPython==3.1.32
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.2
jsonschema==4.18.2
jsonschema-specifications==2023.6.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
mpmath==1.3.0
networkx==3.1
numpy==1.25.1
packaging==23.1
pandas==2.0.3
Pillow==9.5.0
protobuf==4.23.4
pyarrow==12.0.1
pydeck==0.8.1b0
Pygments==2.15.1
Pympler==1.0.1
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
referencing==0.29.1
requests==2.31.0
rich==13.4.2
rpds-py==0.8.10
six==1.16.0
smmap==5.0.0
streamlit==1.24.1
sympy==1.12
tenacity==8.2.2
toml==0.10.2
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tornado==6.3.2
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
urllib3==2.0.3
validators==0.20.0
zipp==3.16.1
altair==5.0.1
attrs==23.1.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.4
comet-ml==3.33.6
configobj==5.0.8
contourpy==1.1.0
cycler==0.11.0
decorator==5.1.1
dulwich==0.21.5
everett==3.1.0
filelock==3.12.2
fonttools==4.41.0
fsspec==2023.6.0
gitdb==4.0.10
GitPython==3.1.32
huggingface-hub==0.16.4
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.2
joblib==1.3.1
jsonschema==4.18.2
jsonschema-specifications==2023.6.1
kiwisolver==1.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
mdurl==0.1.2
mpmath==1.3.0
networkx==3.1
nltk==3.8.1
numpy==1.25.1
opencv-python==4.8.0.74
packaging==23.1
pandas==2.0.3
Pillow==9.5.0
protobuf==4.23.4
psutil==5.9.5
pyarrow==12.0.1
pydeck==0.8.1b0
Pygments==2.15.1
Pympler==1.0.1
pyparsing==3.0.9
python-box==6.1.0
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
referencing==0.29.1
regex==2023.6.3
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.4.2
rpds-py==0.8.10
safetensors==0.3.1
scipy==1.11.1
seaborn==0.12.2
semantic-version==2.10.0
sentry-sdk==1.28.1
simplejson==3.19.1
six==1.16.0
smmap==5.0.0
streamlit==1.24.1
sympy==1.12
tenacity==8.2.2
thop==0.1.1.post2209072238
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
transformers==4.31.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
ultralytics==8.0.135
urllib3==2.0.3
validators==0.20.0
websocket-client==1.3.3
wrapt==1.15.0
wurlitzer==3.0.3
zipp==3.16.1
